refactor: move animation logic into hook
