+++ src/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 
 import "./App.css";
 
-import AnimationManager from './containers/AnimationManager';
+import AnimationManager from './components/AnimationManager';
 
 function App() {
   return (
